Re-org things a bit
We'll have a single index and zip file for this handler and the new one.
